Organize src component files
