--- frontend/src/components/Exercises.js
@@ -1,9 +1,0 @@
-import React, { Component } from 'react';
-
-class Exercises extends Component {
-  render() {
-    return <div>EXERCISE COMPONENT</div>;
-  }
-}
-
-export default Exercises;
