--- frontend/src/components/Games.js
@@ -1,9 +1,0 @@
-import React, { Component } from 'react';
-
-class Games extends Component {
-  render() {
-    return <div>GAME COMPONENT</div>;
-  }
-}
-
-export default Games;
