--- frontend/src/components/Homefeed.js
@@ -1,9 +1,0 @@
-import React, { Component } from 'react';
-
-class Homefeed extends Component {
-  render() {
-    return <div>HOMEFEED COMPONENT</div>;
-  }
-}
-
-export default Homefeed;
