--- frontend/src/components/QuestionList.js
@@ -1,37 +1,0 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
-import Question from './Question';
-
-class QuestionList extends Component {
-  render() {
-    return (
-      <div>
-        {this.props.questions.map((question, i) => {
-          return <Question key={i} {...question} />;
-        })}
-      </div>
-    );
-  }
-}
-
-QuestionList.propTypes = {
-  /* Verify that questions follows the following schema
-    questions = [
-        {
-            title: ...
-            options: ...
-            correctAnswers ...
-        },
-        ...
-    ]
-    */
-  questions: PropTypes.arrayOf(
-    PropTypes.shape({
-      title: PropTypes.string.isRequired,
-      options: PropTypes.arrayOf(PropTypes.string).isRequired,
-      correctAnswer: PropTypes.number.isRequired
-    })
-  ).isRequired
-};
-
-export default QuestionList;
