--- frontend/src/components/Readings.js
@@ -1,9 +1,0 @@
-import React, { Component } from 'react';
-
-class Readings extends Component {
-  render() {
-    return <div>READING COMPONENT</div>;
-  }
-}
-
-export default Readings;
