--- frontend/src/components/Results.js
@@ -1,9 +1,0 @@
-import React, { Component } from 'react';
-
-class Results extends Component {
-  render() {
-    return <div>RESULTS COMPONENT</div>;
-  }
-}
-
-export default Results;
