+++ frontend/src/components/catalog/CatalogCardBook.js
@@ -9,12 +9,11 @@ import {
 } from 'reactstrap';
 import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
-import { URLParamToString } from '../utils/formatHelpers';
+import { URLParamToString } from '../../utils/formatHelpers';
 import { library } from '@fortawesome/fontawesome-svg-core';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faBook } from '@fortawesome/free-solid-svg-icons';
-import '../styles/ReadingOlympics.scss';
-import book from '../images/book-fa.png';
+import '../../styles/ReadingOlympics.scss';
+import book from '../../images/book-fa.png';
 library.add(faBook);
 
 class CatalogCardBook extends Component {
