+++ frontend/src/components/search/RedirectingSearchBar.js
