+++ frontend/src/pages/ReadingOlympics/ROHome.js
@@ -2,7 +2,7 @@ import { Link } from 'react-router-dom';
 import React, { Component } from 'react';
 import { Container, Col, Button, Card, CardBody, CardTitle } from 'reactstrap';
 import '../../styles/ReadingOlympics.scss';
-import RedirectingSearchBar from '../../components/RedirectingSearchBar';
+import RedirectingSearchBar from '../../components/search/RedirectingSearchBar';
 import { URLParamToString } from '../../utils/formatHelpers';
 import { getROArchiveYears } from '../../utils/api';
