+++ frontend/src/pages/RegisterPage.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import Register from '../components/Register';
+import Register from '../components/account/Register';
 
 class RegisterPage extends Component {
   render() {
