+++ frontend/src/pages/admin/AdminEditQuizPage.js
@@ -1,10 +1,10 @@
 import React, { Component } from 'react';
 import { Container, Row, Col, Alert } from 'reactstrap';
-import AdminQuizForm from '../../components/AdminQuizForm';
-import AdminNavigator from '../../components/AdminNavigator';
-import AdminBookSelect from '../../components/AdminBookSelect';
-import AdminQuizSelect from '../../components/AdminQuizSelect';
-import AdminDeleteModal from '../../components/AdminDeleteModal';
+import AdminQuizForm from '../../components/admin/AdminQuizForm';
+import AdminNavigator from '../../components/admin/AdminNavigator';
+import AdminBookSelect from '../../components/admin/AdminBookSelect';
+import AdminQuizSelect from '../../components/admin/AdminQuizSelect';
+import AdminDeleteModal from '../../components/admin/AdminDeleteModal';
 import '../../styles/admin/AdminNavigator.scss';
 import '../../styles/admin/AdminHome.scss';
 import { editQuiz, deleteQuiz } from '../../utils/api.js';
