--- frontend/src/styles/ParentsPage.scss
@@ -1,2 +1,0 @@
-.parents-page {
-}
